+++ source/Public/Capacity/Get-FabricCapacityRefreshables.ps1
@@ -36,5 +36,6 @@ Author: Ioana Bouariu
 
     # Make a GET request to the PowerBI API to retrieve the top refreshable capacities.
     # The function returns the 'value' property of the response.
-    return (Invoke-RestMethod -uri "$($PowerBI.BaseApiUrl)/capacities/refreshables?`$top=$top" -Headers $FabricSession.HeaderParams -Method GET).value
+    $result = Invoke-FabricRestMethod -Method GET -PowerBIApi -Uri "capacities/refreshables?`$top=$top"
+    $result.value
 }