+++ source/Public/Capacity/Suspend-FabricCapacity.ps1
@@ -6,20 +6,20 @@ Suspends a capacity.
 .DESCRIPTION
 The Suspend-FabricCapacity function suspends a capacity. It supports multiple aliases for flexibility.
 
-.PARAMETER subscriptionID
+.PARAMETER SubscriptionID
 The ID of the subscription. This is a mandatory parameter. This is a parameter found in Azure, not Fabric.
 
-.PARAMETER resourcegroup
+.PARAMETER ResourceGroup
 The resource group. This is a mandatory parameter. This is a parameter found in Azure, not Fabric.
 
-.PARAMETER capacity
+.PARAMETER Capacity
 The the capacity. This is a mandatory parameter. This is a parameter found in Azure, not Fabric.
 
 .EXAMPLE
     This example suspends a capacity given the subscription ID, resource group, and capacity.
 
     ```powershell
-    Suspend-FabricCapacity -subscriptionID "your-subscription-id" -resourcegroupID "your-resource-group" -capacityID "your-capacity"
+    Suspend-FabricCapacity -SubscriptionID "your-subscription-id" -ResourceGroup "your-resource-group" -Capacity "your-capacity"
     ```
 
 .NOTES
@@ -35,21 +35,24 @@ Author: Ioana Bouariu
     # Define parameters for the subscription ID, resource group, and capacity.
     Param (
         [Parameter(Mandatory = $true)]
-        [guid]$subscriptionID,
+        [guid]$SubscriptionID,
         [Parameter(Mandatory = $true)]
-        [string]$resourcegroup,
+        [string]$ResourceGroup,
         [Parameter(Mandatory = $true)]
-        [string]$capacity
+        [string]$Capacity
     )
 
     Confirm-TokenState
 
+    $AzureBaseApiUrl = Get-PSFConfigValue 'FabricTools.AzureApi.BaseUrl'
+    $headers = Get-PSFConfigValue 'FabricTools.AzureSession.Headers'
+
     # Define the URI for the request.
-    $suspendCapacity = "$($AzureSession.BaseApiUrl)/subscriptions/$subscriptionID/resourceGroups/$resourcegroup/providers/Microsoft.Fabric/capacities/$capacity/suspend?api-version=2023-11-01"
+    $suspendCapacity = "$AzureBaseApiUrl/subscriptions/$SubscriptionID/resourceGroups/$ResourceGroup/providers/Microsoft.Fabric/capacities/$Capacity/suspend?api-version=2023-11-01"
 
     # Make a GET request to the URI and return the response.
     if ($PSCmdlet.ShouldProcess("Suspend capacity $capacity")) {
-        return Invoke-RestMethod -Method POST -Uri $suspendCapacity -Headers $script:AzureSession.HeaderParams -ErrorAction Stop
+        return Invoke-RestMethod -Method POST -Uri $suspendCapacity -Headers $headers -ErrorAction Stop
     }
 
 }