+++ source/Public/Config/Get-FabricConfig.ps1
@@ -1,23 +1,23 @@
 function Get-FabricConfig {
     <#
 .SYNOPSIS
-Gets the configuration for use with all functions in the PSFabricTools module.
+Gets the configuration for use with all functions in the FabricTools module.
 
 .DESCRIPTION
-Gets the configuration for use with all functions in the PSFabricTools module.
+Gets the configuration for use with all functions in the FabricTools module.
 
 .PARAMETER ConfigName
 The name of the configuration to be retrieved.
 
 .EXAMPLE
-    Gets all configuration values for the PSFabricTools module and outputs them.
+    Gets all configuration values for the FabricTools module and outputs them.
 
     ```powershell
     Get-FabricConfig
     ```
 
 .EXAMPLE
-    Gets the BaseUrl configuration value for the PSFabricTools module.
+    Gets the BaseUrl configuration value for the FabricTools module.
 
     ```powershell
     Get-FabricConfig -ConfigName BaseUrl
@@ -29,12 +29,12 @@ The name of the configuration to be retrieved.
 
 #>
     param (
-        [String]$ConfigName
+        [String] $ConfigName
     )
 
     if ($ConfigName) {
-        Get-PSFConfig -Module PSFabricTools -Name $ConfigName
+        Get-PSFConfig -Module 'FabricTools' -Name $ConfigName
     } else {
-        Get-PSFConfig -Module PSFabricTools
+        Get-PSFConfig -Module 'FabricTools'
     }
 }